Move consts to home
