+++ views/home/components/Certification.js
@@ -2,8 +2,8 @@ import React, { useState } from 'react'
 import styled from 'styled-components/macro'
 import CertItem from '~/views/home/components/certification/CertItem'
 import CertInfo from '~/views/home/components/certification/CertInfo'
-import { CERTIFICATIONS } from '~/consts/pages'
-import { ITEM, INFO } from '~/consts/certification'
+import { CERTIFICATIONS } from '~/views/home/consts/pages'
+import { ITEM, INFO } from '~/views/home/consts/certification'
 import Header from '~/views/home/components/common/section/Header'
 import Anchor from '~/views/home/components/common/Anchor'
 import Row from 'react-bootstrap/Row'
