+++ views/home/components/Navbar.js
@@ -3,7 +3,7 @@ import React, { useState } from 'react'
 import styled from 'styled-components/macro'
 import NavHeader from '~/views/home/components/nav/NavHeader'
 import NavContent from '~/views/home/components/nav/NavContent'
-import { INIT_NAV_CONTENT_SHOW } from '~/consts/initState'
+import { INIT_NAV_CONTENT_SHOW } from '~/views/home/consts/initState'
 
 const Navbar = () => {
   const [isShowNavContent, switchNav] = useState(INIT_NAV_CONTENT_SHOW)
