+++ views/home/components/certification/CertInfo.js
@@ -2,7 +2,7 @@ import React from 'react'
 import styled from 'styled-components/macro'
 import Info from '~/views/home/components/certification/Info'
 import Title from '~/views/home/components/certification/Title'
-import { LEFT_INFO_POS } from '~/consts/certification'
+import { LEFT_INFO_POS } from '~/views/home/consts/certification'
 import PropTypes from 'prop-types'
 import { Translate } from 'react-localize-redux'
 import Col from 'react-bootstrap/Col'
