+++ views/home/components/nav/NavLang.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react'
 import styled from 'styled-components/macro'
-import { INIT_SELECTED_LANG } from '~/consts/initState'
-import { EN, VN } from '~/consts/langs'
+import { INIT_SELECTED_LANG } from '~/views/home/consts/initState'
+import { EN, VN } from '~/views/home/consts/langs'
 import NavLink from '~/views/home/components/nav/NavLink'
 
 const NavLang = () => {
