+++ views/home/components/nav/__test__/NavContent.spec.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import { mountWithTheme, shallowWithTheme } from '~/views/core/utils/withThemeProviders'
 import NavContent from '~/views/home/components/nav/NavContent'
 import * as helpers from '~/views/core/utils/helpers'
-import { ABOUT } from '~/consts/pages'
+import { ABOUT } from '~/views/home/consts/pages'
 
 describe('Nav Content', () => {
   it('should render without throwing an error', function() {
