+++ views/home/consts/certification.js
