+++ views/home/consts/initState.js
@@ -1,5 +1,5 @@
-import { HOME } from '~/consts/pages'
-import { EN } from '~/consts/langs'
+import { HOME } from '~/views/home/consts/pages'
+import { EN } from '~/views/home/consts/langs'
 
 export const INIT_NAV_CONTENT_SHOW = false
 export const INIT_SELECTED_PAGE = HOME
