+++ views/home/consts/typist.js
